requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
babel==2.17.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
deep-translator==1.11.4
deepl==1.23.0
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
docxcompose==1.4.0
docxtpl==0.20.1
filelock
fsspec==2025.9.0
groq==0.32.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
iniconfig
Jinja2==3.1.6
jiter
lxml==6.0.2
MarkupSafe==3.0.3
numpy
odfpy==1.4.1
packaging==25.0
pluggy==1.6.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.26.5
PyPDF2==3.0.1
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pytest==8.4.2
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
vulture==2.14